backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
websockets==12.0
sqlalchemy==2.0.23
alembic==1.13.1

# Security dependencies
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Additional utilities
redis==5.0.1
python-jose[cryptography]==3.3.0
slowapi==0.1.9
bleach==6.1.0
email-validator==2.1.0
python-magic==0.4.27

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
pillow==10.1.0
